+++ src/hooks.tsx
@@ -6,7 +6,6 @@ import {
   useContext,
   MutableRefObject,
   useEffect,
-  DependencyList,
   useRef,
 } from 'react'
 import { LayoutChangeEvent, StyleSheet, ViewProps } from 'react-native'
@@ -21,12 +20,12 @@ import Animated, {
   withDelay,
   withTiming,
   interpolate,
-  Extrapolate,
   runOnJS,
   runOnUI,
-  useDerivedValue,
   useEvent,
   useHandler,
+  AnimatedRef,
+  Extrapolation,
 } from 'react-native-reanimated'
 import { useDeepCompareMemo } from 'use-deep-compare'
 
@@ -38,7 +37,6 @@ import {
   TabName,
   TabReactElement,
   TabsWithProps,
-  Ref,
 } from './types'
 
 export function useContainerRef() {
@@ -47,19 +45,15 @@ export function useContainerRef() {
 
 export function useAnimatedDynamicRefs(): [
   ContextType['refMap'],
-  ContextType['setRef']
+  ContextType['setRef'],
 ] {
   const [map, setMap] = useState<ContextType['refMap']>({})
-  const setRef = useCallback(function <T extends RefComponent>(
-    key: TabName,
-    ref: React.RefObject<T>
-  ) {
+  const setRef = useCallback((key: TabName, ref: AnimatedRef<RefComponent>) => {
     setMap((map) => ({ ...map, [key]: ref }))
     return ref
-  },
-  [])
+  }, [])
 
-  return [map, setRef]
+  return [map, setRef as ContextType['setRef']]
 }
 
 export function useTabProps<T extends TabName>(
@@ -124,6 +118,20 @@ export function useTabNameContext(): TabName {
   return c
 }
 
+export function useLayoutHeight(initialHeight: number = 0) {
+  const [height, setHeight] = useState(initialHeight)
+
+  const getHeight = useCallback(
+    (event: LayoutChangeEvent) => {
+      const latestHeight = event.nativeEvent.layout.height
+      if (latestHeight !== height) {
+        setHeight(latestHeight)
+      }
+    },
+    [height, setHeight]
+  )
+  return [height, getHeight] as const
+}
 /**
  * Hook to access some key styles that make the whole thing work.
  *
@@ -138,15 +146,9 @@ export function useCollapsibleStyle(): CollapsibleStyle {
     allowHeaderOverscroll,
     minHeaderHeight,
   } = useTabsContext()
-  const [containerHeightVal, tabBarHeightVal, headerHeightVal] = [
-    useConvertAnimatedToValue(containerHeight),
-    useConvertAnimatedToValue(tabBarHeight),
-    useConvertAnimatedToValue(headerHeight),
-  ]
-
   const containerHeightWithMinHeader = Math.max(
     0,
-    (containerHeightVal ?? 0) - minHeaderHeight
+    (containerHeight ?? 0) - minHeaderHeight
   )
 
   return useMemo(
@@ -155,24 +157,24 @@ export function useCollapsibleStyle(): CollapsibleStyle {
       contentContainerStyle: {
         minHeight:
           IS_IOS && !allowHeaderOverscroll
-            ? containerHeightWithMinHeader - (tabBarHeightVal || 0)
-            : containerHeightWithMinHeader + (headerHeightVal || 0),
+            ? containerHeightWithMinHeader - (tabBarHeight || 0)
+            : containerHeightWithMinHeader + (headerHeight || 0),
         paddingTop:
           IS_IOS && !allowHeaderOverscroll
             ? 0
-            : (headerHeightVal || 0) + (tabBarHeightVal || 0),
+            : (headerHeight || 0) + (tabBarHeight || 0),
       },
       progressViewOffset:
         // on iOS we need the refresh control to be at the top if overscrolling
         IS_IOS && allowHeaderOverscroll
           ? 0
           : // on android we need it below the header or it doesn't show because of z-index
-            (headerHeightVal || 0) + (tabBarHeightVal || 0),
+            (headerHeight || 0) + (tabBarHeight || 0),
     }),
     [
       allowHeaderOverscroll,
-      headerHeightVal,
-      tabBarHeightVal,
+      headerHeight,
+      tabBarHeight,
       width,
       containerHeightWithMinHeader,
     ]
@@ -226,7 +228,7 @@ export function useScroller<T extends RefComponent>() {
 
   const scroller = useCallback(
     (
-      ref: Ref<T> | undefined,
+      ref: AnimatedRef<T> | undefined,
       x: number,
       y: number,
       animated: boolean,
@@ -236,9 +238,9 @@ export function useScroller<T extends RefComponent>() {
       if (!ref) return
       //! this is left here on purpose to ease troubleshooting (uncomment when necessary)
       // console.log(
-      //   `${_debugKey}, y: ${y}, y adjusted: ${y - contentInset.value}`
+      //   `${_debugKey}, y: ${y}, y adjusted: ${y - contentInset}`
       // )
-      scrollToImpl(ref, x, y - contentInset.value, animated)
+      scrollToImpl(ref, x, y - contentInset, animated)
     },
     [contentInset]
   )
@@ -254,7 +256,6 @@ export const useScrollHandlerY = (name: TabName) => {
     revealHeaderOnScroll,
     refMap,
     tabNames,
-    index,
     headerHeight,
     contentInset,
     containerHeight,
@@ -272,12 +273,14 @@ export const useScrollHandlerY = (name: TabName) => {
 
   const enabled = useSharedValue(false)
 
+  const scrollTo = useScroller()
+
   const enable = useCallback(
     (toggle: boolean) => {
       'worklet'
       enabled.value = toggle
     },
-    [enabled]
+    [name, refMap, scrollTo]
   )
 
   /**
@@ -287,13 +290,6 @@ export const useScrollHandlerY = (name: TabName) => {
    * call it to sync the scenes.
    */
   const afterDrag = useSharedValue(0)
-
-  const tabIndex = useMemo(() => tabNames.value.findIndex((n) => n === name), [
-    tabNames,
-    name,
-  ])
-
-  const scrollTo = useScroller()
 
   const scrollAnimation = useSharedValue<number | undefined>(undefined)
 
@@ -358,11 +354,6 @@ export const useScrollHandlerY = (name: TabName) => {
     }
   }
 
-  const contentHeight = useDerivedValue(() => {
-    const tabIndex = tabNames.value.indexOf(name)
-    return contentHeights.value[tabIndex] || Number.MAX_VALUE
-  }, [])
-
   const scrollHandler = useAnimatedScrollHandler(
     {
       onScroll: (event) => {
@@ -372,23 +363,31 @@ export const useScrollHandlerY = (name: TabName) => {
           if (IS_IOS) {
             let { y } = event.contentOffset
             // normalize the value so it starts at 0
-            y = y + contentInset.value
+            y = y + contentInset
+
+            const contentHeight =
+              contentHeights.value[tabNames.value.indexOf(name)] ||
+              Number.MAX_VALUE
+
             const clampMax =
-              contentHeight.value -
-              (containerHeight.value || 0) +
-              contentInset.value
+              contentHeight - (containerHeight || 0) + contentInset
             // make sure the y value is clamped to the scrollable size (clamps overscrolling)
             scrollYCurrent.value = allowHeaderOverscroll
               ? y
-              : interpolate(y, [0, clampMax], [0, clampMax], Extrapolate.CLAMP)
+              : interpolate(
+                  y,
+                  [0, clampMax],
+                  [0, clampMax],
+                  Extrapolation.CLAMP
+                )
           } else {
             const { y } = event.contentOffset
             scrollYCurrent.value = y
           }
 
-          scrollY.value[index.value] = scrollYCurrent.value
+          scrollY.value[name] = scrollYCurrent.value
           oldAccScrollY.value = accScrollY.value
-          accScrollY.value = scrollY.value[index.value] + offset.value
+          accScrollY.value = scrollY.value[name] + offset.value
 
           if (revealHeaderOnScroll) {
             const delta = accScrollY.value - oldAccScrollY.value
@@ -472,8 +471,8 @@ export const useScrollHandlerY = (name: TabName) => {
         focusedTab.value !== name
       ) {
         let nextPosition: number | null = null
-        const focusedScrollY = scrollY.value[Math.round(indexDecimal.value)]
-        const tabScrollY = scrollY.value[tabIndex]
+        const focusedScrollY = scrollY.value[focusedTab.value]
+        const tabScrollY = scrollY.value[name]
         const areEqual = focusedScrollY === tabScrollY
 
         if (!areEqual) {
@@ -492,7 +491,7 @@ export const useScrollHandlerY = (name: TabName) => {
             if (focusedIsOnTop) {
               nextPosition = snappingTo.value
             } else if (currIsOnTop) {
-              nextPosition = headerHeight.value || 0
+              nextPosition = headerHeight || 0
             }
           } else if (currIsOnTop || focusedIsOnTop) {
             nextPosition = Math.min(focusedScrollY, headerScrollDistance.value)
@@ -501,12 +500,12 @@ export const useScrollHandlerY = (name: TabName) => {
 
         if (nextPosition !== null) {
           // console.log(`sync ${name} ${nextPosition}`)
-          scrollY.value[tabIndex] = nextPosition
+          scrollY.value[name] = nextPosition
           scrollTo(refMap[name], 0, nextPosition, false, `[${name}] sync pane`)
         }
       }
     },
-    [revealHeaderOnScroll, refMap, snapThreshold, tabIndex, enabled, scrollTo]
+    [revealHeaderOnScroll, refMap, snapThreshold, enabled, scrollTo]
   )
 
   return { scrollHandler, enable }
@@ -547,20 +546,8 @@ export function useAfterMountEffect(
   nextOnLayout: ViewProps['onLayout'],
   effect: React.EffectCallback
 ) {
-  const name = useTabNameContext()
-  const {
-    //tabsMounted,
-    refMap,
-    scrollY,
-    //scrollYCurrent,
-    tabNames,
-  } = useTabsContext()
-
   const didExecute = useRef(false)
   const didMount = useSharedValue(false)
-
-  const scrollTo = useScroller()
-  const ref = name ? refMap[name] : null
 
   useAnimatedReaction(
     () => {
@@ -569,16 +556,7 @@ export function useAfterMountEffect(
     (didMount, prevDidMount) => {
       if (didMount && !prevDidMount) {
         if (didExecute.current) return
-        if (ref) {
-          const tabIndex = tabNames.value.findIndex((n) => n === name)
-          scrollTo(
-            ref,
-            0,
-            scrollY.value[tabIndex],
-            false,
-            `[${name}] restore scroll position`
-          )
-        }
+
         effect()
         didExecute.current = true
       }
@@ -601,7 +579,7 @@ export function useAfterMountEffect(
 export function useConvertAnimatedToValue<T>(
   animatedValue: Animated.SharedValue<T>
 ) {
-  const [value, setValue] = useState(animatedValue.value)
+  const [value, setValue] = useState<T>(animatedValue.value)
 
   useAnimatedReaction(
     () => {
@@ -615,7 +593,7 @@ export function useConvertAnimatedToValue<T>(
     [value]
   )
 
-  return value
+  return value || 0
 }
 
 export interface HeaderMeasurements {
@@ -626,7 +604,7 @@ export interface HeaderMeasurements {
   /**
    * Animated value that represents the height of the header
    */
-  height: Animated.SharedValue<number | undefined>
+  height: number
 }
 
 export function useHeaderMeasurements(): HeaderMeasurements {
@@ -669,7 +647,7 @@ export const usePageScrollHandler = (
       context: unknown
     ) => unknown
   },
-  dependencies?: DependencyList
+  dependencies?: unknown[]
 ) => {
   const { context, doDependenciesDiffer } = useHandler(handlers, dependencies)
   const subscribeForEvents = ['onPageScroll']